Rename macros header file
