+++ source/tests/system/nirfmxbluetooth_driver_api_tests.cpp
@@ -6,7 +6,7 @@
 #include "nirfmxbluetooth/nirfmxbluetooth_client.h"
 #include "nirfmxbluetooth/nirfmxbluetooth_service.h"
 #include "nirfsa/nirfsa_client.h"
-#include "rfmx_expect_macros.h"
+#include "rfmx_macros.h"
 #include "waveform_helpers.h"
 
 namespace pb = google::protobuf;
