+++ source/tests/system/nirfmxspecan_driver_api_tests.cpp
@@ -8,7 +8,7 @@
 #include "nirfmxinstr/nirfmxinstr_client.h"
 #include "nirfmxspecan/nirfmxspecan_client.h"
 #include "nirfmxspecan/nirfmxspecan_service.h"
-#include "rfmx_expect_macros.h"
+#include "rfmx_macros.h"
 #include "waveform_helpers.h"
 
 using namespace nirfmxspecan_grpc;
